refactor: split agentic report family
